Commit #007
Coyote now copies and loads natives on its own, increasing
cross-platform abilities. As things are added to Coyote, more natives
will need to be loaded and will be added in as needed. Natives are
technically loaded on CoreGame initialization, but that should be
initialized in every application. Also made Coyote a little more
verbose.